added conditions to restrict functions to logged in user

+++ view/forum/topics/displayTopic.php
@@ -28,23 +28,27 @@
         (<?= $topic->getCreationDate() ?>)
     </p>
 
-    <div>
-        <a href="index.php?ctrl=topic&action=displayModTopicForm&id=<?= $topic->getID() ?>">
-            <i class="fa-solid fa-pen-to-square"></i> Edit topic
-        </a>
-    </div>
+    <?php
+    if(App\Session::getUser() && App\Session::getUser() == $topic->getUser()){ ?>
+        <div>
+            <a href="index.php?ctrl=topic&action=displayModTopicForm&id=<?= $topic->getID() ?>">
+                <i class="fa-solid fa-pen-to-square"></i> Edit topic
+            </a>
+        </div>
 
-    <div>
-        <a href="index.php?ctrl=topic&action=closeTopic&id=<?= $topic->getID() ?>">
-            <i class="fa-solid fa-square-xmark"></i> Close topic
-        </a>
-    </div>
+        <div>
+            <a href="index.php?ctrl=topic&action=closeTopic&id=<?= $topic->getID() ?>">
+                <i class="fa-solid fa-square-xmark"></i> Close topic
+            </a>
+        </div>
 
-    <div>
-        <a href="index.php?ctrl=topic&action=deleteTopic&id=<?= $topic->getID() ?>">
-            <i class="fa-solid fa-trash"></i> Delete topic (forever)
-        </a>
-    </div>
+        <div>
+            <a href="index.php?ctrl=topic&action=deleteTopic&id=<?= $topic->getID() ?>">
+                <i class="fa-solid fa-trash"></i> Delete topic (forever)
+            </a>
+        </div>
+    <?php } ?>
+    
 </div>
 
 <div id="topicReplies">
@@ -70,11 +74,13 @@
                     (<?= $post->getCreationDate() ?>)
                 </p>
 
+                <?php
+                if(App\Session::getUser() && App\Session::getUser() == $post->getUser()){ ?>
+
                 <div>
                     <a href="index.php?ctrl=topic&action=displayModPostForm&id=<?= $post->getID() ?>">
                         <i class="fa-solid fa-pen-to-square"></i> Edit reply
                     </a>
-                    
                 </div>
 
                 <div>
@@ -82,6 +88,8 @@
                         <i class="fa-solid fa-trash"></i> Delete reply (forever)
                     </a>
                 </div>
+
+                <?php } ?>
             </div>
 
     <?php }
@@ -101,17 +109,20 @@ if($topic->getClosed() == 1){ ?>
 <?php }
 else { ?>
 
-<div>  
-    <form action="index.php?ctrl=topic&action=addPost&id=<?= $topic->getID() ?>" method="post">
+    <?php
+    if(App\Session::getUser()){ ?>
+    <div>  
+        <form action="index.php?ctrl=topic&action=addPost&id=<?= $topic->getID() ?>" method="post">
 
-        <div>
-            <label for="contentInput">Your reply : </label>
-            <textarea name="inputContent" id="contentInput" cols="30" rows="10" placeholder="Enter text here"></textarea>
-        </div>
+            <div>
+                <label for="contentInput">Your reply : </label>
+                <textarea name="inputContent" id="contentInput" cols="30" rows="10" placeholder="Enter text here"></textarea>
+            </div>
 
-        <input type="submit" name="submit" id="postSubmit" value="Post">
+            <input type="submit" name="submit" id="postSubmit" value="Post">
 
-    </form>
-</div>
+        </form>
+    </div>
+    <?php } ?>
 
 <?php } ?>
